+++ tests/benchmarks/benchmarks.py
@@ -108,7 +108,6 @@ def test_tile(nodata_type, data_type, dataset_name, dataset_fixture, benchmark):
                 nodata_type=nodata_type,
                 width=4000,
                 height=4000,
-                filled=True,
             )
         ) as memfile:
             with memfile.open() as dst: